Remove unused packages.config file, clean up using directives, and improve code readability across multiple files

+++ src/Libraries/ACATExtension/UI/AlphabetScannerCommon.cs
@@ -415,7 +415,7 @@ namespace ACAT.Lib.Extension
 
                 form.Top = 0;
                 form.Height = desktopHeight;
-                form.Width = (int)((float)form.Width * ratio);
+                form.Width = (int)(form.Width * ratio);
             }
         }
